+++ resources/views/default/column/editable/checkbox.blade.php
@@ -17,7 +17,7 @@
   @endif
 
   {!! $append !!}
-  
+
   @if($small)
     <small class="clearfix">{!! $small !!}</small>
   @endif