+++ resources/views/default/column/editable/datetime.blade.php
@@ -11,6 +11,7 @@
       data-mode="{{ $mode }}"
       data-combodate="{{ $combodateValue  }}"
       data-pk="{{ $id }}"
+      data-emptytext="{{ trans('sleeping_owl::lang.select.no_items') }}"
       data-disabled="{{ !$isEditable }}"
   >{{ $text }}</a>