+++ resources/views/default/column/editable/textarea.blade.php
@@ -7,6 +7,7 @@
       data-type="textarea"
       data-pk="{{ $id }}"
       data-mode="{{ $mode }}"
+      data-emptytext="{{ trans('sleeping_owl::lang.select.no_items') }}"
       data-disabled="{{ !$isEditable }}"
   >{{ $text }}</a>